refactor(frontend): OrganizationCategoriesEdit: OrganizationCategoriesEditPage delegates logic to OrganizationCategoriesEdit component
using new component "OrganizationCategoriesEdit" for showing and adding categories logic on OrganizationCategoriesEditPage
